Remove FunctionDefinition
