--- src/FunctionDefinitionInterface.php
@@ -1,9 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace webignition\BasilCompilationSource;
-
-interface FunctionDefinitionInterface extends LineListInterface
-{
-    public function getName(): string;
-    public function getArguments(): array;
-}
